Moving form to containers folder.
It is connected to the store, so it is not a dumb component.
